+++ src/components/ThemeProvider/index.tsx
@@ -1,4 +1,4 @@
-// ThemeProvider from copy-ui @ 2025-10-08.
+// ThemeProvider from copy-ui @ 2025-10-10.
 
 export type { ColorName, ColorNo } from "./ThemeProvider";
 export { ThemeContext, ThemeProvider, useTheme } from "./ThemeProvider";